+++ src/Symfony/Bundle/FrameworkBundle/Resources/views/Form/date_widget.html.php
@@ -1,5 +1,5 @@
 <?php if ($widget == 'single_text'): ?>
-    <?php echo $view['form']->renderBlock('form_widget_single_control'); ?>
+    <?php echo $view['form']->renderBlock('form_widget_simple'); ?>
 <?php else: ?>
     <div <?php echo $view['form']->renderBlock('widget_container_attributes') ?>>
         <?php echo str_replace(array('{{ year }}', '{{ month }}', '{{ day }}'), array(