+++ src/Symfony/Bundle/FrameworkBundle/Resources/views/Form/field_widget.html.php
@@ -1,1 +1,1 @@
-<?php echo $view['form']->renderBlock('form_widget_single_control') ?>
+<?php echo $view['form']->renderBlock('form_widget_simple') ?>