+++ src/Symfony/Bundle/FrameworkBundle/Resources/views/Form/form_row.html.php
@@ -1,6 +1,6 @@
 <div>
     <?php echo $view['form']->label($form, isset($label) ? $label : null) ?>
-    <?php if ($single_control): ?>
+    <?php if (!$compound): ?>
         <?php echo $view['form']->errors($form) ?>
     <?php endif ?>
     <?php echo $view['form']->widget($form) ?>