+++ src/Symfony/Component/Form/Extension/Core/Type/DateType.php
@@ -165,8 +165,8 @@ class DateType extends AbstractType
      */
     public function setDefaultOptions(OptionsResolverInterface $resolver)
     {
-        $singleControl = function (Options $options) {
-            return $options['widget'] === 'single_text';
+        $compound = function (Options $options) {
+            return $options['widget'] !== 'single_text';
         };
 
         $resolver->setDefaults(array(
@@ -188,7 +188,7 @@ class DateType extends AbstractType
             // representation is not \DateTime, but an array, we need to unset
             // this option.
             'data_class'     => null,
-            'single_control' => $singleControl,
+            'compound'       => $compound,
         ));
 
         $resolver->setAllowedValues(array(