+++ src/Symfony/Component/Form/Extension/Core/Type/MoneyType.php
@@ -51,11 +51,11 @@ class MoneyType extends AbstractType
     public function setDefaultOptions(OptionsResolverInterface $resolver)
     {
         $resolver->setDefaults(array(
-            'precision'      => 2,
-            'grouping'       => false,
-            'divisor'        => 1,
-            'currency'       => 'EUR',
-            'single_control' => true,
+            'precision' => 2,
+            'grouping'  => false,
+            'divisor'   => 1,
+            'currency'  => 'EUR',
+            'compound'  => false,
         ));
     }