+++ src/Symfony/Component/Form/Extension/Core/Type/NumberType.php
@@ -37,10 +37,10 @@ class NumberType extends AbstractType
     {
         $resolver->setDefaults(array(
             // default precision is locale specific (usually around 3)
-            'precision'      => null,
-            'grouping'       => false,
-            'rounding_mode'  => \NumberFormatter::ROUND_HALFUP,
-            'single_control' => true,
+            'precision'     => null,
+            'grouping'      => false,
+            'rounding_mode' => \NumberFormatter::ROUND_HALFUP,
+            'compound'      => false,
         ));
 
         $resolver->setAllowedValues(array(