+++ src/Symfony/Component/Form/Extension/Core/Type/TextType.php
@@ -34,7 +34,7 @@ class TextType extends AbstractType
     public function setDefaultOptions(OptionsResolverInterface $resolver)
     {
         $resolver->setDefaults(array(
-            'single_control' => true,
+            'compound' => false,
         ));
     }